refactor(RadioGroupChildren): decompose RadioGroupChildren
